alx_travel_app/requirement.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
gunicorn==23.0.0
sqlparse==0.5.3
uritemplate==4.2.0
typing_extensions==4.14.0

